requirements.txt
anyio==3.7.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
backcall==0.2.0
catboost==1.0.6
category-encoders==2.5.0
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.1.8
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.75.1
fonttools==4.57.0
graphviz==0.20.3
h11==0.16.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
loguru==0.6.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
narwhals==1.41.0
nest-asyncio==1.6.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
plotly==6.1.2
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==1.9.1
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pyzmq==26.4.0
requests==2.27.1
scikit-learn==1.1.1
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.35
stack-data==0.6.3
starlette==0.17.1
statsmodels==0.14.1
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.16.0
wcwidth==0.2.13
zipp==3.20.2
